update calendar view

+++ backend/src/main/java/cn/edu/shiep/backend/meetingroom/dto/request/ReservationRequest.java
@@ -1,11 +1,8 @@
 package cn.edu.shiep.backend.meetingroom.dto.request;
 
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
 @Data
@@ -15,5 +12,5 @@ public class ReservationRequest {
     private int personNum;
     private LocalDateTime startTime;
     private LocalDateTime endTime;
-    private List<Integer> equipmentIds;
+    private List<Long> equipmentIds;
 }
